requirements.txt
Django==1.7.1
Markdown==2.5.2
PyYAML==3.11
Pygments==2.0.1
South==1.0.2
defusedxml==0.4.1
django-bootstrap3==5.0.3
django-filter==0.9.1
django-guardian==1.2.5
django-markdown-deux==1.0.5
django-markdown2==0.2.1
django-oauth-plus==2.2.5
django-oauth2-provider==0.2.6.1
djangorestframework==3.0.2
httplib2==0.9
markdown2==2.3.0
mock==1.0.1
oauth2==1.5.211
shortuuid==0.4.2
six==1.8.0
wsgiref==0.1.2
